Removed vk, wh, register, and roles commands
