+++ src/discord/discord_database_access.py
@@ -97,7 +97,7 @@ def check_for_value(column, value_check):
     conn = get_db_connection()
     cur = conn.cursor()
     cur.execute(f"SELECT Discord FROM User WHERE {column} = ?", [value_check])
-    item = cur.fetchone()[0]
+    item = cur.fetchone()
     close_db_connection(conn)
     if item:
         return item
